backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
cmake==3.27.0
deepface==0.0.96
tensorflow==2.20.0